debugged the default case being called on removeTodo()
Earlier when the void main was calling the removeTodo() function, the user had to input the number of the todo he/she wanted to remove. But after the users input the default case of the switch case was being called, to enter only one of 'add, remove, display or quit'. So I have just editted the code such that, that the void main will ignore the default case when removeTodo() is called. **I have not debugged it just have found out a way to hide the bug.